Allow setting cursor position outside of visible area of the view
It is allowed to set it to a position outside the visible portion of the view;
this is needed if the view is scrolled and the cursor position is in the portion
of the view that is scrolled off-screen.

It is even allowed to set the position to invalid values with respect to the
view content (e.g. less than zero, or y greater than the length of the viewlines
array). It might seem reasonable to clamp to valid values, but we can't do that
because the methods are often called at a time where the viewlines array is not
up to date. It is the caller's responsibility to clamp to valid values.

Note that errors in client code that gets this wrong are harmless. The cursor
position is only used to draw the cursor; if it is out of bounds, the worst that
will happen is that the cursor blinks in the wrong place. We don't have to be
worried about out-of-bounds access to the viewlines array, for example; it is
not used for that.