Don't draw cursor for editable views when outside of visible area
This improves the situation where you have an editable view whose content is
taller than the screen, and you scroll it with the mouse wheel so that the
cursor leaves the visible area of the view. Previously we would clamp the cursor
position to the visible area, which means we would effectively move the cursor
in the opposite direction when you scroll. (But only visually; once you then
type a character, the view would scroll to show the real cursor position, and
the visual cursor would jump back to its real position.)

Also, fix an off-by-two error where the cursor would still draw when it is on
the lower frame of the view, or one line below; the reason is that we were using
Size() rather than InnerSize() to determine the valid visible area.