Fix typing at the end of a very tall editable view
We had an off-by-one error here where if the cursor is equal to the size of the
view, we would leave it there. That's already outside of the visible area, so we
need to scroll the view up by one line in that case to make the cursor visible.

In lazygit this would happen in commit descriptions that are so long that they
show a scrollbar.