Add IsDoubleClick field to ViewMouseBindingOpts
A click is considered to be a double click if
- it is within 500ms from the previous click
- it is in the same view as the previous click
- it is using the same mouse button as the previous click
- it is on the same view-local coordinates as the previous one.

The last condition is important for the case where the first click changes the
view geometry; this can happen in lazygit if the expandFocusedSidePanel config
is on. If you click in an inactive side panel, this will change the geometry of
the views, so the clicked view will probably move up or down a bit; if you then
quickly click again, you'll hit a different line index than with the first
click, and it would be bad to treat this as a double click.