Allow typing unprintable characters into TextArea
This is needed for being able to "type" emojis such as 🏴󠁧󠁢󠁥󠁮󠁧󠁿; such an emoji is a
grapheme cluster that consists of a base rune (🏴 in this case), and a bunch of
tag sequences that are not printable when they are used on their own.

It is not really possible to type such an emoji (as far as I can tell), but you
can copy/paste it, in which case the individual runes of this grapheme cluster
will be passed to SimpleEditor one by one, and it is important that we add them
all to the text area's content string.